cli: extract runtime functions
